added fmt test, that checks the order of comments remains unchanged

+++ tests_fmt/test.sh
@@ -34,14 +34,17 @@ find "$TEST_DIR" -name '*.fut' | xargs -P $THREADS -I {} sh -c '
     futhark fmt "$prog" 2> /dev/null > "$name.fmt.fut"
     futhark fmt "$name.fmt.fut" 2> /dev/null > "$name.fmt.fmt.fut"
     futhark hash "$name.fmt.fut" 2> /dev/null > "$prog.actual"
+    futhark tokens "$prog" 2> /dev/null | grep '^COMMENT' > "$name.comments"
+    futhark tokens "$name.fmt.fut" 2> /dev/null | grep '^COMMENT' > "$name.fmt.comments"
 
     hash_result=1
     idempotent_result=1
+    comments_result=1
 
     if ! cmp --silent "$prog.expected" "$prog.actual"
     then
         hash_result=0
-        echo "Failed hash Comparison Test" >> "$name.log"
+        echo "Failed Hash Comparison Test" >> "$name.log"
     fi
     printf "$hash_result" >> "$name.result"
 
@@ -51,12 +54,21 @@ find "$TEST_DIR" -name '*.fut' | xargs -P $THREADS -I {} sh -c '
         echo "Failed Idempotent Comparison Test" >> "$name.log"
     fi
     printf "$idempotent_result" >> "$name.result"
+
+    if ! cmp --silent "$name.comments" "$name.fmt.comments"
+    then
+        comments_result=0
+        echo "Failed Order of Comments Test" >> "$name.log"
+    fi 
+    printf "$comments_result" >> "$name.result"
 '
 
 idempotent_pass=0
 idempotent_fail=0
 hash_pass=0
 hash_fail=0
+comments_pass=0
+comments_fail=0
 
 for file in $(find "$TEST_DIR" -name '*.result')
 do
@@ -67,13 +79,16 @@ do
         content=$(cat "$file")
         hash_result=$(printf "$content" | cut -c1)
         idempotent_result=$(printf "$content" | cut -c2)
+        comments_result=$(printf "$content" | cut -c3)
 
         hash_pass=$((hash_pass + hash_result))
         hash_fail=$((hash_fail + 1 - hash_result))
         idempotent_pass=$((idempotent_pass + idempotent_result))
         idempotent_fail=$((idempotent_fail + 1 - idempotent_result))
+        comments_pass=$((comments_pass + comments_result))
+        comments_faul=$((comments_pass + 1 - comments_result))
 
-        if [ "$hash_result" -eq 1 ] && [ "$idempotent_result" -eq 1 ]; then
+        if [ "$hash_result" -eq 1 ] && [ "$idempotent_result" -eq 1 ] && [ "$comments_result" -eq 1 ]; then
             rm "${file%.*}."*
         fi
     fi
@@ -83,3 +98,4 @@ find "$TEST_DIR" -type d -empty -delete
 
 echo "Hash Tests Passed: $hash_pass/$((hash_pass + hash_fail))"
 echo "Idempotent Tests Passed: $idempotent_pass/$((idempotent_pass + idempotent_fail))"
+echo "Order of Comments Tests Passed: $comments_pass/$((comments_pass + comments_fail))"
